+++ backend/PyMatcha/utils/tasks.py
@@ -1,5 +1,4 @@
 import datetime
-import json
 import logging
 from math import ceil
 
@@ -7,10 +6,7 @@ from PyMatcha import celery
 from PyMatcha import redis
 from PyMatcha.models.message import Message
 from PyMatcha.models.user import User
-from PyMatcha.utils.match_score import _get_age_diff
-from PyMatcha.utils.match_score import _get_common_tags
-from PyMatcha.utils.match_score import _get_distance
-from PyMatcha.utils.match_score import _get_gender_query
+from PyMatcha.utils.recommendations import create_user_recommendations
 
 BASE_HEAT_SCORE = 30
 LIKES_MULTIPLIER = 2
@@ -102,66 +98,11 @@ def update_offline_users():
     )
 
 
-def default_date_converter(o):
-    if isinstance(o, datetime.datetime):
-        return o.__str__()
-
-
 @celery.task
 def update_user_recommendations():
-    today = datetime.datetime.utcnow()
     count = 0
     for user_to_update in User.select_all():
-        count += 1
-        user_to_update_recommendations = []
-        if not user_to_update.birthdate:
-            continue
-        if not user_to_update.geohash:
-            continue
-
-        user_to_update_age = (
-            today.year
-            - user_to_update.birthdate.year
-            - ((today.month, today.day) < (user_to_update.birthdate.month, user_to_update.birthdate.day))
-        )
-        user_to_update_tags = [t.name for t in user_to_update.get_tags()]
-
-        query = _get_gender_query(user_to_update.orientation, user_to_update.gender)
-
-        for user in query:
-            if user.id == user_to_update.id:
-                continue
-            score = 0
-
-            distance = _get_distance(user_to_update.geohash, user.geohash)
-            if distance:
-                score -= distance
-
-            user_age = (
-                today.year
-                - user.birthdate.year
-                - ((today.month, today.day) < (user.birthdate.month, user.birthdate.day))
-            )
-            age_diff = _get_age_diff(user_to_update_age, user_age)
-            score -= age_diff
-
-            user_tags = [t.name for t in user.get_tags()]
-            common_tags = _get_common_tags(user_to_update_tags, user_tags)
-            score += len(common_tags) * 2
-
-            score += user.heat_score
-
-            d = {"score": score, "common_tags": common_tags, "distance": distance}
-            d.update(user.to_dict())
-            user_to_update_recommendations.append(d)
-        user_to_update_recommendations_sorted = sorted(
-            user_to_update_recommendations, key=lambda x: x["score"], reverse=True
-        )
-        redis.set(
-            f"user_recommendations:{str(user_to_update.id)}",
-            json.dumps(user_to_update_recommendations_sorted, default=default_date_converter),
-        )
-        redis.expire(f"user_recommendations:{str(user_to_update.id)}", 3600)
+        create_user_recommendations(user_to_update)
     return f"Successfully updated recommendations for {count} users."
